# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bff_elearning",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "PAYLOAD_TOO_LARGE",
                  "UNPROCESSABLE_ENTITY",
                  "TOO_MANY_REQUESTS",
                  "INTERNAL_ERROR",
                  "BAD_GATEWAY",
                  "SERVICE_UNAVAILABLE",
                  "GATEWAY_TIMEOUT"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/elearning/admin/courses": {
      "post": {
        "tags": [
          "E-learning administration"
        ],
        "summary": "Creates a course (not available yet: 501)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCourseCreateBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "course": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "accueil-agents"
                        },
                        "title": {
                          "type": "string",
                          "example": "Accueil des nouveaux agents"
                        },
                        "description": {
                          "type": "string"
                        },
                        "instructor": {
                          "type": "string"
                        },
                        "rating": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "string"
                            }
                          ],
                          "example": 4.8
                        },
                        "duration": {
                          "type": "string",
                          "example": "2 h 15"
                        },
                        "chapters": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "string"
                            }
                          ],
                          "example": 4
                        },
                        "learners": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "string"
                            }
                          ],
                          "example": 128
                        },
                        "category": {
                          "type": "string",
                          "example": "Integration"
                        },
                        "statusValue": {
                          "type": "string",
                          "enum": [
                            "not-started",
                            "in-progress",
                            "completed"
                          ],
                          "description": "Statut de progression de la formation pour l’utilisateur connecté",
                          "example": "in-progress"
                        },
                        "titleBadge": {
                          "type": "object",
                          "properties": {
                            "label": {
                              "type": "string",
                              "example": "Obligatoire"
                            },
                            "variant": {
                              "type": "string",
                              "enum": [
                                "default",
                                "beginner",
                                "intermediate",
                                "advanced",
                                "inProgress",
                                "completed",
                                "mandatory",
                                "notStarted"
                              ],
                              "description": "Variante visuelle utilisée par les badges de formation",
                              "example": "mandatory"
                            }
                          },
                          "required": [
                            "label"
                          ],
                          "description": "Badge associé à une formation"
                        },
                        "levelBadge": {
                          "type": "object",
                          "properties": {
                            "label": {
                              "type": "string",
                              "example": "Obligatoire"
                            },
                            "variant": {
                              "type": "string",
                              "enum": [
                                "default",
                                "beginner",
                                "intermediate",
                                "advanced",
                                "inProgress",
                                "completed",
                                "mandatory",
                                "notStarted"
                              ],
                              "description": "Variante visuelle utilisée par les badges de formation",
                              "example": "mandatory"
                            }
                          },
                          "required": [
                            "label"
                          ],
                          "description": "Badge associé à une formation"
                        },
                        "statusBadge": {
                          "type": "object",
                          "properties": {
                            "label": {
                              "type": "string",
                              "example": "Obligatoire"
                            },
                            "variant": {
                              "type": "string",
                              "enum": [
                                "default",
                                "beginner",
                                "intermediate",
                                "advanced",
                                "inProgress",
                                "completed",
                                "mandatory",
                                "notStarted"
                              ],
                              "description": "Variante visuelle utilisée par les badges de formation",
                              "example": "mandatory"
                            }
                          },
                          "required": [
                            "label"
                          ],
                          "description": "Badge associé à une formation"
                        },
                        "progress": {
                          "type": "number",
                          "minimum": 0,
                          "maximum": 100
                        },
                        "deadline": {
                          "type": "string",
                          "example": "30 juin 2026"
                        },
                        "ratingDistribution": {
                          "type": "object",
                          "properties": {
                            "1": {
                              "type": "number",
                              "minimum": 0
                            },
                            "2": {
                              "type": "number",
                              "minimum": 0
                            },
                            "3": {
                              "type": "number",
                              "minimum": 0
                            },
                            "4": {
                              "type": "number",
                              "minimum": 0
                            },
                            "5": {
                              "type": "number",
                              "minimum": 0
                            }
                          },
                          "description": "Répartition des notes par valeur, de 1 à 5",
                          "example": {
                            "1": 0,
                            "2": 1,
                            "3": 4,
                            "4": 35,
                            "5": 88
                          }
                        },
                        "details": {
                          "type": "object",
                          "properties": {
                            "title": {
                              "type": "string"
                            },
                            "subtitle": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string"
                            },
                            "instructor": {
                              "type": "string"
                            },
                            "duration": {
                              "type": "string"
                            },
                            "rating": {
                              "anyOf": [
                                {
                                  "type": "number"
                                },
                                {
                                  "type": "string"
                                }
                              ]
                            },
                            "ratingLabel": {
                              "type": "string"
                            },
                            "ratingDistribution": {
                              "type": "object",
                              "properties": {
                                "1": {
                                  "type": "number",
                                  "minimum": 0
                                },
                                "2": {
                                  "type": "number",
                                  "minimum": 0
                                },
                                "3": {
                                  "type": "number",
                                  "minimum": 0
                                },
                                "4": {
                                  "type": "number",
                                  "minimum": 0
                                },
                                "5": {
                                  "type": "number",
                                  "minimum": 0
                                }
                              },
                              "description": "Répartition des notes par valeur, de 1 à 5",
                              "example": {
                                "1": 0,
                                "2": 1,
                                "3": 4,
                                "4": 35,
                                "5": 88
                              }
                            },
                            "progress": {
                              "type": "number",
                              "minimum": 0,
                              "maximum": 100
                            },
                            "completed": {
                              "type": "boolean"
                            },
                            "completionRating": {
                              "type": "object",
                              "properties": {
                                "initialValue": {
                                  "type": "number",
                                  "minimum": 1,
                                  "maximum": 5
                                },
                                "max": {
                                  "type": "number",
                                  "minimum": 1
                                },
                                "submitted": {
                                  "type": "boolean"
                                },
                                "disabled": {
                                  "type": "boolean"
                                },
                                "title": {
                                  "type": "string"
                                },
                                "helperText": {
                                  "type": "string"
                                },
                                "submitLabel": {
                                  "type": "string"
                                },
                                "submittedLabel": {
                                  "type": "string"
                                }
                              },
                              "description": "Configuration du composant de notation de fin de formation"
                            },
                            "chapters": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string",
                                    "example": "accueil-1"
                                  },
                                  "title": {
                                    "type": "string",
                                    "example": "Municipal organisation"
                                  },
                                  "description": {
                                    "type": "string"
                                  },
                                  "duration": {
                                    "type": "string",
                                    "example": "25 min"
                                  },
                                  "completed": {
                                    "type": "boolean"
                                  },
                                  "active": {
                                    "type": "boolean"
                                  },
                                  "contents": {
                                    "type": "array",
                                    "items": {
                                      "type": "object",
                                      "properties": {
                                        "id": {
                                          "type": "string",
                                          "example": "accueil-1-video"
                                        },
                                        "title": {
                                          "type": "string",
                                          "example": "Présentation des directions et services"
                                        },
                                        "type": {
                                          "type": "string",
                                          "enum": [
                                            "video",
                                            "pdf",
                                            "document",
                                            "link",
                                            "quiz",
                                            "audio",
                                            "other"
                                          ],
                                          "description": "Type de contenu pédagogique",
                                          "example": "video"
                                        },
                                        "description": {
                                          "type": "string"
                                        },
                                        "duration": {
                                          "type": "string",
                                          "example": "12 min"
                                        },
                                        "fileName": {
                                          "type": "string",
                                          "example": "organigramme-mairie360.pdf"
                                        },
                                        "href": {
                                          "type": "string",
                                          "example": "/documents/organigramme-mairie360.pdf"
                                        },
                                        "completed": {
                                          "type": "boolean"
                                        },
                                        "required": {
                                          "type": "boolean"
                                        }
                                      },
                                      "required": [
                                        "id",
                                        "title",
                                        "type"
                                      ],
                                      "description": "Contenu pédagogique d’un chapitre"
                                    }
                                  }
                                },
                                "required": [
                                  "id",
                                  "title"
                                ],
                                "description": "Chapitre d’une formation"
                              }
                            }
                          },
                          "required": [
                            "title",
                            "description",
                            "chapters"
                          ],
                          "description": "Détail complet affiché dans la modale de formation"
                        }
                      },
                      "required": [
                        "id",
                        "title",
                        "description"
                      ],
                      "description": "Formation affichée dans le catalogue"
                    }
                  },
                  "required": [
                    "course"
                  ],
                  "description": "Formation créée ou mise à jour par un administrateur"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Restricted to administrators",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not available: no upstream service stores this data yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "BFF User or an upstream API is unreachable, not configured, failed or answered an unusable body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}